[FTR] remove SDL and game field and walls new renderer

+++ Manager/VboIndexer.cpp
@@ -2,7 +2,7 @@
 
 namespace Manager {
     void VboIndexer::indexVBO(vector<glm::vec3> &in_vertices, vector<glm::vec2> &in_uvs, vector<glm::vec3> &in_normals,
-                              vector<unsigned short> &out_indices, vector<glm::vec3> &out_vertices,
+                              vector<unsigned int> &out_indices, vector<glm::vec3> &out_vertices,
                               vector<glm::vec2> &out_uvs, vector<glm::vec3> &out_normals) {
         std::map<PackedVertex,unsigned short> VertexToOutIndex;
 
@@ -32,7 +32,7 @@ namespace Manager {
     void
     VboIndexer::indexVBO_TBN(vector<glm::vec3> &in_vertices, vector<glm::vec2> &in_uvs, vector<glm::vec3> &in_normals,
                              vector<glm::vec3> &in_tangents, vector<glm::vec3> &in_bitangents,
-                             vector<unsigned short> &out_indices, vector<glm::vec3> &out_vertices,
+                             vector<unsigned int> &out_indices, vector<glm::vec3> &out_vertices,
                              vector<glm::vec2> &out_uvs, vector<glm::vec3> &out_normals,
                              vector<glm::vec3> &out_tangents, vector<glm::vec3> &out_bitangents) {
         // For each input vertex
